pyproject.toml
[tool.poetry]
name = "automacao-operacional"
version = "0.1.0"
description = ""
authors = ["lucasrluz <lucasr.luzbr@gmail.com>"]
readme = "README.md"
packages = [{include = "automacao_operacional"}]

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.7.2"
playwright = "^1.29.1"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
